test(interaction): split Telegram suites into test/unit/interaction/plugins/telegram.test.ts
interaction-plugins.test.ts was at 790 of the 800-line test limit, leaving no
room for the authorization regression tests. Pure move -- no test body changes.
Destination follows the existing test/unit/interaction/plugins/ convention
already established by cli.test.ts.
